requirements.txt
Django>=4.2,<5.0
python-decouple>=3.8
Pillow>=10.0.0
whitenoise>=6.5.0
gunicorn>=21.2.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
boto3>=1.26.0
django-storages>=1.13.0

